// client/src/app/shared/model/comparision.model.ts
// Generated by https://quicktype.io

export interface ComparisionSettings {
  config: Config;
  id: number;
  name: string;
  type: string;
  date: string;
  active: boolean;
}

export interface Config {
  matching_peaks: string;
  matching_score: string;
  pep_mass_tolerance: number[];
}

export interface NewComparisonSettings {
  name: string;
  matching_peaks: string;
  matching_score: string;
  pepmass_tolerance_from: string | number;
  pepmass_tolerance_to: string | number;
}